+++ packages/lit-element/src/lit-element.ts
@@ -55,7 +55,7 @@
  * @packageDocumentation
  */
 import {PropertyValues, UpdatingElement} from 'updating-element';
-import {render, RenderOptions, noChange} from 'lit-html';
+import {render, RenderOptions, noChange, NodePart} from 'lit-html';
 export * from 'updating-element';
 export * from 'lit-html';
 
@@ -97,6 +97,8 @@ export class LitElement extends UpdatingElement {
 
   readonly _renderOptions: RenderOptions = {eventContext: this};
 
+  private _nodePart: NodePart | undefined = undefined;
+
   protected createRenderRoot() {
     const renderRoot = super.createRenderRoot();
     // When adoptedStyleSheets are shimmed, they are inserted into the
@@ -120,7 +122,17 @@ export class LitElement extends UpdatingElement {
     // before that.
     const value = this.render();
     super.update(changedProperties);
-    render(value, this.renderRoot, this._renderOptions);
+    this._nodePart = render(value, this.renderRoot, this._renderOptions);
+  }
+
+  connectedCallback() {
+    super.connectedCallback();
+    this._nodePart?.setDirectiveConnection(true);
+  }
+
+  disconnectedCallback() {
+    super.connectedCallback();
+    this._nodePart?.setDirectiveConnection(false);
   }
 
   /**