+++ packages/updating-element/src/updating-element.ts
@@ -524,6 +524,9 @@ export abstract class UpdatingElement extends HTMLElement {
   // connected before first update.
   private _updatePromise!: Promise<unknown>;
 
+  private _pendingConnectionPromise: Promise<unknown> | undefined = undefined;
+  private _enableConnection: (() => void) | undefined = undefined;
+
   isUpdatePending = false;
   hasUpdated = false;
 
@@ -620,6 +623,12 @@ export abstract class UpdatingElement extends HTMLElement {
     }
     this.enableUpdating();
     this._controllers?.forEach((c) => c.connectedCallback?.());
+    // If we were disconnected, re-enable updating by resolving the pending
+    // connection promise
+    if (this._enableConnection) {
+      this._enableConnection();
+      this._pendingConnectionPromise = this._enableConnection = undefined;
+    }
   }
 
   /**
@@ -636,6 +645,9 @@ export abstract class UpdatingElement extends HTMLElement {
    */
   disconnectedCallback() {
     this._controllers?.forEach((c) => c.disconnectedCallback?.());
+    this._pendingConnectionPromise = new Promise(
+      (r) => (this._enableConnection = r)
+    );
   }
 
   /**
@@ -767,6 +779,10 @@ export abstract class UpdatingElement extends HTMLElement {
       // Ensure any previous update has resolved before updating.
       // This `await` also ensures that property changes are batched.
       await this._updatePromise;
+      // If we were disconnected, wait until re-connected to flush an update
+      while (this._pendingConnectionPromise) {
+        await this._pendingConnectionPromise;
+      }
     } catch (e) {
       // Refire any previous errors async so they do not disrupt the update
       // cycle. Errors are refired so developers have a chance to observe